refactor: move stubs to separed files
