+++ test/stubs/hash-stub.ts
@@ -1,0 +1,19 @@
+import { IHasher, IHashCompare } from '@/use-cases/interfaces';
+
+export const makeHasher = (): IHasher => {
+  class HasherStub implements IHasher {
+    async hash(value: string): Promise<string> {
+      return 'any_hash';
+    }
+  }
+  return new HasherStub();
+};
+
+export const makeHashCompare = (): IHashCompare => {
+  class HashCompareStub implements IHashCompare {
+    async compare(hash: string, value: string): Promise<boolean> {
+      return true;
+    }
+  }
+  return new HashCompareStub();
+};
