+++ test/stubs/id-generator-stub.ts
@@ -1,0 +1,10 @@
+import { IIdGenerator } from '@/use-cases/interfaces';
+
+export const makeIdGenerator = (): IIdGenerator => {
+  class IdGeneratorStub implements IIdGenerator {
+    async generate(): Promise<string> {
+      return 'any_id';
+    }
+  }
+  return new IdGeneratorStub();
+};
