+++ test/stubs/use-cases-stub.ts
@@ -1,0 +1,39 @@
+import { SignUpUseCase } from '@/use-cases/sign-up';
+import { SignInUseCase } from '@/use-cases/sign-in';
+import { AuthenticationUseCase } from '@/use-cases/authentication';
+import { UpdateUserUseCase } from '@/use-cases/update-user';
+import {
+  makeUserRepository,
+  makeHasher,
+  makeIdGenerator,
+  makeEncrypter,
+  makeHashCompare,
+  makeDecrypter,
+} from '@/test/stubs';
+
+export const makeSignUpUseCase = (): SignUpUseCase => {
+  const userRepository = makeUserRepository();
+  const hasher = makeHasher();
+  const idGenerator = makeIdGenerator();
+  const encrypter = makeEncrypter();
+  return new SignUpUseCase(userRepository, hasher, idGenerator, encrypter);
+};
+
+export const makeSignInUseCase = (): SignInUseCase => {
+  const userRepository = makeUserRepository();
+  const hashCompare = makeHashCompare();
+  const encrypter = makeEncrypter();
+  return new SignInUseCase(userRepository, hashCompare, encrypter);
+};
+
+export const makeAuthenticationUseCase = (): AuthenticationUseCase => {
+  const userRepository = makeUserRepository();
+  const decrypter = makeDecrypter();
+  return new AuthenticationUseCase(userRepository, decrypter);
+};
+
+export const makeUpdateUserUseCase = (): UpdateUserUseCase => {
+  const userRepository = makeUserRepository();
+  const hasher = makeHasher();
+  return new UpdateUserUseCase(userRepository, hasher);
+};
