+++ test/stubs/validation-stub.ts
@@ -1,0 +1,10 @@
+import { IValidation } from '@/adapters/interfaces';
+
+export const makeValidation = (): IValidation => {
+  class ValidationStub implements IValidation {
+    validate(input: any): Error | null {
+      return null;
+    }
+  }
+  return new ValidationStub();
+};
